make some fkn changes

+++ db.py
@@ -8,36 +8,33 @@ API_URL = os.getenv("API_URL")
 API_KEY = os.getenv("API_KEY")
 
 if not API_URL or not API_KEY:
-    raise Exception(
-        "API_URL or API_KEY not found in .env file")
+    raise Exception("API_URL or API_KEY not found in .env file")
 
 supabase = create_client(API_URL, API_KEY)
 
+
 def get_users():
-    return (
-        supabase.table("Users")
-        .select("*")
-        .execute()
-        .data
-    )
+    return supabase.table("Users").select("*").execute().data
 
 
 def get_user(user_id: int):
-    return (
-        supabase.table("Users")
-        .select("*")
-        .eq("id", user_id)
-        .execute()
-        .data[0]
-    )
+    return supabase.table("Users").select("*").eq("id", user_id).execute().data[0]
+
+
+def get_user_by_email(email: str):
+    users = supabase.table("Users").select("*").eq("email", email).execute().data
+    if len(users) == 0:
+        return None
+    else:
+        return users[0]
 
 
 def add_user(
     username: str = "",
     name: str = "",
-    email: str= "",
-    google_refresh_token: str= "",
-    terra_user_id: str = " "
+    email: str = "",
+    google_refresh_token: str = "",
+    terra_user_id: str = " ",
 ):
     return (
         supabase.table("Users")
@@ -47,7 +44,7 @@ def add_user(
                 "name": name,
                 "email": email,
                 "google_refresh_token": google_refresh_token,
-                "terra_user_id": terra_user_id
+                "terra_user_id": terra_user_id,
             }
         )
         .execute()
@@ -58,9 +55,9 @@ def edit_user(
     user_id: int,
     username: str = "",
     name: str = "",
-    email: str= "",
-    google_refresh_token: str= "",
-    terra_user_id: str = ""
+    email: str = "",
+    google_refresh_token: str = "",
+    terra_user_id: str = "",
 ):
     data = {}
 
@@ -72,37 +69,22 @@ def edit_user(
 
     if email != "":
         data["email"] = email
-    
+
     if terra_user_id != "":
         data["terra_user_id"] = terra_user_id
 
     if google_refresh_token != "":
         data["google_refresh_token"] = google_refresh_token
 
-    return (
-        supabase.table("Users")
-        .update(data)
-        .eq("id", user_id)
-        .execute()
-    )
+    return supabase.table("Users").update(data).eq("id", user_id).execute()
 
 
 def delete_user(user_id: int):
-    return (
-        supabase.table("Users")
-        .delete()
-        .eq("id", user_id)
-        .execute()
-    )
+    return supabase.table("Users").delete().eq("id", user_id).execute()
 
 
 def get_tasks():
-    return (
-        supabase.table("Tasks")
-        .select("*")
-        .execute()
-        .data
-    )
+    return supabase.table("Tasks").select("*").execute().data
 
 
 def get_tasks_from_user(user_id: int):
@@ -119,10 +101,14 @@ def get_tasks_from_user(user_id: int):
 def add_task(user_id: int, name: str, description: str = "", recurring: bool = False):
     return (
         supabase.table("Tasks")
-        .insert({"user_id": user_id,
-                 "name": name,
-                 "description": description,
-                 "recurring": recurring})
+        .insert(
+            {
+                "user_id": user_id,
+                "name": name,
+                "description": description,
+                "recurring": recurring,
+            }
+        )
         .execute()
     )
 
@@ -136,21 +122,11 @@ def edit_task(task_id: int, name: str, description: str = ""):
     if description != "":
         data["description"] = description
 
-    return (
-        supabase.table("Tasks")
-        .update(data)
-        .eq("id", task_id)
-        .execute()
-    )
+    return supabase.table("Tasks").update(data).eq("id", task_id).execute()
 
 
 def delete_task(task_id: int):
-    return (
-        supabase.table("Tasks")
-        .delete()
-        .eq("id", task_id)
-        .execute()
-    )
+    return supabase.table("Tasks").delete().eq("id", task_id).execute()
 
 
 def mark_task_as_added(task_id):
